provider: Update cert request on property change (#21)
It used to be the case that once you made a certificate request; if you
changed the resoruce in puppet, the cert request wouldn't be updated. It
relied on using the force_resubmit flag always, which is not ideal and
wasn't the intention. This is no longer the case with this change.

The force_resubmit flag should work when one doesn't actually change
properties from the resource, but still wants to issue a resubmit.
Currently we don't have that capability, but it will be added in the
future.

For now, making changes to the resource now works.

Change-Id: I2a682199299732adf31e9aed008ad6488d6f011c