Don't quote pre/post save commands
Commit cf86dd14b6e9c2de692fcae3ffd22b8c9f9c7a02 introduced quoting for
these parameters. However, the quoting ended up being double.

The arguments are passed to the command as a list, so the quoting was
not needed as one item in the list is taken as a single parameter.

This quoting broke post/pre save commands, so we gotta remove it to make
it work again.

Change-Id: I284f72d438ed00a7c25de51f2e1ea4f04395c83b