Fix comparison logic for key_size and eku
eku is an array of values, which can be in any order.  The changes
in this patch update the insync? method to reflect this so that
spurious changes are not triggered.

key_size was added to allow callers to specify a key_size different
from the default.  Unfortunately, getcert list does not provide
the key_size output.  This results in a failing comparison every time.

To fix this, a mechanism is added to obtain the key_size by checking
the key file with openssl.